Update : LottoController 분리
